USSD_APP/requirements.txt
asgiref==3.9.0
Django==5.2.4
django-rest-framework==0.1.0
djangorestframework==3.16.1
sqlparse==0.5.3
tzdata==2025.2
